Added done/undone toggle + weather api + counter + other features and styling

+++ main.js
@@ -1,12 +1,20 @@
-let taskInput = document.getElementById("task");
+const taskInput = document.getElementById("task");
 const btnAdd = document.getElementById("add");
 const btnUpdate = document.getElementById("update");
 const btnDelete = document.getElementById("delete");
-let counter;
+const deleteList = document.getElementById("clearList");
+const taskStatus = document.getElementById("display");
+const taskUndone = document.getElementsByClassName("fa-circle");
+const taskDone = document.getElementsByClassName("fa-circle-check");
+const counter = document.getElementById("counter");
+const weather = document.getElementById("weather");
+let task = document.getElementById("display");
 let taskIndex;
+let myWeather = [];
 
 let taskContainer = JSON.parse(localStorage.getItem("Task")) || [];
 displayTodo();
+displayWeather();
 
 btnAdd.addEventListener("click", addTodo);
 
@@ -21,10 +29,38 @@ taskInput.addEventListener("keyup", function (event) {
 });
 btnUpdate.addEventListener("click", updateTask);
 
+deleteList.addEventListener("click", clearList);
+
+//toggling between done and undone
+taskStatus.addEventListener("click", function (event) {
+  const taskRow = event.target.closest("tr");
+  taskIndex = Array.from(taskRow.parentElement.children).indexOf(taskRow);
+  const taskText = taskRow.querySelector(".taskText");
+
+  if (event.target.classList.contains("fa-circle")) {
+    event.target.classList.add("d-none");
+    event.target.nextElementSibling.classList.remove("d-none");
+    taskText.classList.add("done");
+    taskContainer[taskIndex].taskStatus = "done";
+  } else if (event.target.classList.contains("fa-circle-check")) {
+    event.target.classList.add("d-none");
+    event.target.previousElementSibling.classList.remove("d-none");
+    taskText.classList.remove("done");
+
+    taskContainer[taskIndex].taskStatus = "undone";
+  }
+  localStorage.setItem("Task", JSON.stringify(taskContainer));
+  footerCounter();
+});
+
 function addTodo() {
   let taskInputValue = taskInput.value;
   if (taskInputValue) {
-    taskContainer.push(taskInputValue);
+    let taskObject = {
+      taskItem: taskInputValue,
+      taskStatus: "undone",
+    };
+    taskContainer.push(taskObject);
     localStorage.setItem("Task", JSON.stringify(taskContainer));
     displayTodo();
     clearInputs();
@@ -40,15 +76,21 @@ function clearInputs() {
 function displayTodo() {
   let loopContainer = "";
   if (taskContainer.length == 0) {
-    loopContainer += `<td class="fw-light text-center">Task List is Empty!</td>`;
+    loopContainer += `<td colspan=4 class=" fw-light text-center"> Your task list is empty!</td>`;
   } else {
     for (let i = 0; i < taskContainer.length; i++) {
       loopContainer += `
                 <tr>
-                    <td>${i + 1}</td>
                     <td>
-                    <span>${taskContainer[i]}</span>
+                      <i style="cursor: pointer" class="fa-regular fa-circle"></i>
+                      <i class="fa-regular fa-circle-check d-none"></i>
                     </td>
+                    <td>
+                      <span class="taskText ${
+                        taskContainer[i].taskStatus === "done" ? "done" : ""
+                      }">
+                      ${taskContainer[i].taskItem}
+                      </span> </td>
                     <td>
                     <a style="cursor: pointer" class="update"
                         ><i class="fas fa-pen fa-lg text-warning me-3" onclick="editTaskForm(${i})"></i
@@ -63,11 +105,19 @@ function displayTodo() {
     }
   }
   document.getElementById("display").innerHTML = loopContainer;
+  footerCounter();
+}
+
+function editTaskForm(index) {
+  taskInput.value = taskContainer[index].taskItem;
+  btnUpdate.classList.remove("d-none");
+  btnAdd.classList.add("d-none");
+  taskIndex = index;
 }
 
 function updateTask() {
   if (taskInput.value) {
-    taskContainer[taskIndex] = taskInput.value;
+    taskContainer[taskIndex].taskItem = taskInput.value;
     localStorage.setItem("Task", JSON.stringify(taskContainer));
     displayTodo();
     clearInputs();
@@ -78,31 +128,58 @@ function updateTask() {
   btnAdd.classList.remove("d-none");
 }
 
-function editTaskForm(index) {
-  taskInput.value = taskContainer[index];
-  btnUpdate.classList.remove("d-none");
-  btnAdd.classList.add("d-none");
-  taskIndex = index;
+function deleteTask(index) {
+  Swal.fire({
+    title: "Delete task?",
+    text: "This can't be undone",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Yes, delete it!",
+  }).then((result) => {
+    if (result.isConfirmed) {
+      Swal.fire({
+        title: "Deleted!",
+        text: "Your task has been deleted.",
+        icon: "success",
+      });
+      taskContainer.splice([index], 1);
+      localStorage.setItem("Task", JSON.stringify(taskContainer));
+      displayTodo();
+    }
+  });
 }
 
-function deleteTask(index) {
-  taskContainer.splice([index], 1);
-  localStorage.setItem("Task", JSON.stringify(taskContainer));
+function clearList() {
+  localStorage.clear();
+  taskContainer = [];
   displayTodo();
 }
 
 function footerCounter() {
-  counter = taskContainer.length + 1;
-
-  document.getElementById("counter").innerHTML = counter;
+  let undoneCount = 0;
+  for (let i = 0; i < taskContainer.length; i++) {
+    if (taskContainer[i].taskStatus == "undone") {
+      undoneCount++;
+    }
+  }
+  document.getElementById("counter").innerHTML = undoneCount;
 }
 
-function clearList() {}
-
-
-// replace serial with checkmark
-// dynamic footer counter
-// categorize
-// animations
-// strikethrough done items
-// clearList()
+//Weather API
+async function displayWeather() {
+  try {
+    const response = await fetch(
+      `https://api.weatherapi.com/v1/forecast.json?key=8c3f90441c2c488f919212256241212&q=cairo`
+    );
+    const weatherData = await response.json();
+    document.getElementById("weather").innerHTML = `
+      <div class="h2 poppins-bold text-white text-center">${weatherData.current.temp_c} <sup>o</sup> C</div>`;
+  } catch (error) {
+    console.error(error);
+    document.getElementById(
+      "weather"
+    ).innerHTML = `<div class="text-danger">Unable to fetch weather data.</div>`;
+  }
+}
